add current logout

+++ controllers/contactsController.js
@@ -4,8 +4,11 @@ const { cntrWrap } = require('../decorator');
 
 
 const getAll = async (req, res,) => {
-
-    const result = await Contact.find();
+    const { _id: owner } = req.user
+    const { page = 1, limit = 10, ...qyery } = req.query
+    const skip = (page - 1) * limit
+    const result = await Contact.find({ owner, ...qyery }, "-createdAt -updatedAt -owner", { skip, limit })
+        .populate("owner", "email");
 
     res.json(result);
 
@@ -13,9 +16,9 @@ const getAll = async (req, res,) => {
 
 const getContactById = async (req, res) => {
     const { id } = req.params;
-    console.log(id);
+    const { _id: owner } = req.user
     // const result = await Contact.findOne({ _id: id });
-    const result = await Contact.findById(id);
+    const result = await Contact.findById({ id, owner });
     if (!result) {
 
         throw createNewError(404);
@@ -25,7 +28,9 @@ const getContactById = async (req, res) => {
 
 const updateContactById = async (req, res) => {
     const { id } = req.params;
-    const result = await Contact.findOneAndUpdate({ _id: id }, req.body, { new: true })
+    const { _id: owner } = req.user
+
+    const result = await Contact.findOneAndUpdate({ _id: id, owner }, req.body, { new: true })
     if (!result) {
         throw createNewError(404)
     }
@@ -33,14 +38,15 @@ const updateContactById = async (req, res) => {
 }
 
 const addContact = async (req, res) => {
-
-    const result = await Contact.create(req.body)
+    const { _id: owner } = req.user
+    const result = await Contact.create({ ...req.body, owner })
     res.status(201).json(result)
 }
 
 const deleteContactById = async (req, res) => {
     const { id } = req.params;
-    const result = await Contact.findOneAndDelete({ _id: id });
+    const { _id: owner } = req.user
+    const result = await Contact.findOneAndDelete({ _id: id, owner });
     if (!result) {
         throw createNewError(404);
     }
@@ -48,7 +54,8 @@ const deleteContactById = async (req, res) => {
 }
 const updateStatusContact = async (req, res) => {
     const { id } = req.params;
-    const result = await Contact.findOneAndUpdate({ _id: id }, req.body, { new: true })
+    const { _id: owner } = req.user
+    const result = await Contact.findOneAndUpdate({ _id: id, owner }, req.body, { new: true })
     if (!result) {
         throw createNewError(404)
     }
